+++ lib/domain/model/entity.dart
@@ -554,17 +554,64 @@ class ConceptEntity<E extends ConceptEntity<E>> implements EntityApi {
   int compareTo(E entity) {
     if (code != null) {
       return _code.compareTo(entity.code);
-    } else {
-      if (id == null || entity.id == null) {
-        if (_concept == null) {
-          throw new ConceptError('Entity concept is not defined.');
-        }
-        String msg =
-            '${_concept.code} concept does not have an id.';
-        throw new IdError(msg);
-      }
+    } else if (id != null && entity.id != null) {
       return id.compareTo(entity.id);
+    } else if (!concept.attributes.isEmpty) {
+      return compareAttributes(entity);
+    } else {
+      String msg = '${_concept.code} concept does not have attributes.';
+      throw new IdError(msg);
     }
+  }
+  
+  /**
+   * Compares two entities based on their attributes.
+   * If the result is less than 0 then the first id is less than the second,
+   * if it is equal to 0 they are equal and
+   * if the result is greater than 0 then the first is greater than the second.
+   */
+  int compareAttributes(E entity) {
+    var compare = 0;
+    for (Attribute a in concept.attributes) {
+      if (a.type.base == 'String') {
+        String attribute1 = _attributeMap[a.code];
+        String attribute2 = entity.getAttribute(a.code);
+        compare = attribute1.compareTo(attribute2);
+      } else if (a.type.base == 'num' ||
+        a.type.base == 'int' || a.type.base == 'double') {
+        num attribute1 = _attributeMap[a.code];
+        num attribute2 = entity.getAttribute(a.code);
+        compare = attribute1.compareTo(attribute2);
+      } else if (a.type.base == 'bool') {
+        bool attribute1 = _attributeMap[a.code];
+        bool attribute2 = entity.getAttribute(a.code);
+        compare = attribute1.toString().compareTo(attribute2.toString());
+      } else if (a.type.base == 'DateTime') {
+        DateTime attribute1 = _attributeMap[a.code];
+        DateTime attribute2 = entity.getAttribute(a.code);
+        compare = attribute1.compareTo(attribute2);
+      } else if (a.type.base == 'Duration') {
+        Duration attribute1 = _attributeMap[a.code];
+        Duration attribute2 = entity.getAttribute(a.code);
+        compare = attribute1.compareTo(attribute2);
+      } else if (a.type.base == 'Uri') {
+        Uri attribute1 = _attributeMap[a.code];
+        Uri attribute2 = entity.getAttribute(a.code);
+        compare = attribute1.toString().compareTo(attribute2.toString());
+      } else {
+        String msg = 'cannot compare then order on this type';
+        if (a.concept != null) {
+          msg = '${a.concept.code}.${a.code} is of ${a.type.code} type: cannot order.';
+        } else {
+          msg = '${a.code} is of ${a.type.code} type: cannot order.';
+        }
+        throw new OrderError(msg);
+      }
+      if (compare != 0) {
+        break;
+      }
+    } // for
+    return compare;
   }
 
   /**